refactor: rename cwl file
