+++ packages/core/src/awsService/cloudWatchLogs/commands/viewLogStream.ts
@@ -11,7 +11,7 @@ import * as picker from '../../../shared/ui/picker'
 import { MultiStepWizard, WIZARD_RETRY, WIZARD_TERMINATE, WizardStep } from '../../../shared/wizards/multiStepWizard'
 import { LogGroupNode } from '../explorer/logGroupNode'
 import * as CloudWatchLogsV3 from '@aws-sdk/client-cloudwatch-logs'
-import { CloudWatchLogsClient } from '../../../shared/clients/cloudWatchLogsClient'
+import { CloudWatchLogsClient } from '../../../shared/clients/cloudWatchLogs'
 import { getPaginatedAwsCallIter, IteratorTransformer } from '../../../shared/utilities/collectionUtils'
 import {
     CloudWatchLogsGroupInfo,
