+++ packages/core/src/test/awsService/apprunner/explorer/apprunnerServiceNode.test.ts
@@ -9,7 +9,7 @@ import { AppRunner } from 'aws-sdk'
 import { AppRunnerNode } from '../../../../awsService/apprunner/explorer/apprunnerNode'
 import { AppRunnerServiceNode } from '../../../../awsService/apprunner/explorer/apprunnerServiceNode'
 import { DefaultAppRunnerClient } from '../../../../shared/clients/apprunnerClient'
-import { CloudWatchLogsClient } from '../../../../shared/clients/cloudWatchLogsClient'
+import { CloudWatchLogsClient } from '../../../../shared/clients/cloudWatchLogs'
 import { asyncGenerator } from '../../../../shared/utilities/collectionUtils'
 import { AWSTreeNodeBase } from '../../../../shared/treeview/nodes/awsTreeNodeBase'
 import { stub } from '../../../utilities/stubber'
